+++ src/state/use-beta-account.js
@@ -43,8 +43,6 @@ export const useBetaAccount = () => {
     window.location = window.location.href.split("?")[0];
   };
 
-  console.log("****", `${window.location.origin}?accountID=${state.accountID}`);
-
   return {
     ...state,
     uname,