+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - Import AccountID from URL parameter (step toward sharing the account with another device)
 - Improve layouting and Account guarding
 - Share account via QR Code
+- Copy share url to clipboard